+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v0.4 (TBD)
 --------------------
  - Random number generator functions: `vrand` and `frand`
  - Optionally set DICOM prefix and entropy source in "Advanced"
+ - Keep cross-file UID connections by updating "Referenced" tags
 
 v0.3 (2020.03.31)
 --------------------